course_recommender_web/requirements.txt
amqp==5.0.1
anyjson==0.3.3
asgiref==3.2.10
beautifulsoup4==4.9.1
billiard==3.6.3.0
bs4==0.0.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
dj-database-url==0.5.0
Django==3.1
django-common-helpers==0.9.2
django-cron==0.5.1
django-email-verification==0.0.7
django-floppyforms==1.9.0
django-heroku==0.3.1
django-materializecss-form==1.1.17
django-pandas==0.6.2
django-scheduler==0.9.3
django-timezone-field==4.0
gunicorn==20.0.4
icalendar==4.0.7
idna==2.10
joblib==0.16.0
kombu==5.0.2
numpy==1.19.1
pandas==1.1.1
prompt-toolkit==3.0.7
psycopg2==2.8.6
pycparser==2.20
python-crontab==2.5.1
python-dateutil==2.8.1
pytz==2020.1
rabbitmq==0.2.0
requests==2.24.0
schedule==0.6.0
scikit-surprise==1.1.1
scipy==1.5.2
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
urllib3==1.25.10
vine==5.0.0
wcwidth==0.2.5
whitenoise==5.2.0
